Implemented Journal entries Authentication.

+++ src/main/java/net/backend/journalApp/service/JournalEntryService.java
@@ -29,7 +29,7 @@ public class JournalEntryService {
             JournalEntry saved=journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
             //user.setUserName(userName);
-            userService.saveNewEntry(user);
+            userService.saveUser(user);
         }
         catch (Exception e){
             throw new RuntimeException("An error occurred while saving the entry...",e);
@@ -50,13 +50,26 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id,String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveNewEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id,String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed=user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+        return removed;
+
     }
 
+    public List<JournalEntry> findByUserName(String userName){
+        return null;
+    }
 
 
 }
